+++ pyquil/quil.py
@@ -43,6 +43,7 @@ import numpy as np
 
 from qcs_sdk.compiler.quilc import NativeQuilMetadata
 
+from pyquil.control_flow_graph import ControlFlowGraph
 from pyquil.gates import MEASURE, RESET
 from pyquil.noise import _check_kraus_ops, _create_kraus_pragmas, pauli_kraus_map
 from pyquil.quilatom import (
@@ -319,6 +320,10 @@ class Program:
                     raise ValueError("Invalid instruction: {}, type: {}".format(instruction, type(instruction)))
 
         return self
+
+    def control_flow_graph(self) -> ControlFlowGraph:
+        """Return the :py:class:`~pyquil.control_flow_graph.ControlFlowGraph` of the program."""
+        return ControlFlowGraph._from_rs(self._program.control_flow_graph())
 
     def with_loop(
         self,
@@ -832,7 +837,9 @@ class Program:
         if calibrations:
             return self._program.to_quil()
         else:
-            return self._program.into_simplified().to_quil()
+            return self.filter_instructions(
+                lambda inst: not isinstance(inst, (DefCalibration, DefMeasureCalibration))
+            ).out()
 
     @deprecated(
         version="4.0",
@@ -1012,7 +1019,7 @@ class Program:
 
     def __eq__(self, other: object) -> bool:
         if isinstance(other, Program):
-            return self._program.to_instructions() == other._program.to_instructions()
+            return self._program == other._program
         return False
 
     def __len__(self) -> int:
@@ -1032,6 +1039,12 @@ class Program:
         your program contains unaddressed QubitPlaceholders
         """
         return self._program.to_quil_or_debug()
+
+    def get_all_instructions(self) -> List[AbstractInstruction]:
+        """
+        Get _all_ instructions that makeup the program.
+        """
+        return _convert_to_py_instructions(self._program.to_instructions())
 
 
 def merge_with_pauli_noise(