+++ pyquil/simulation/__init__.py
@@ -1,0 +1,21 @@
+__all__ = [
+    "get_measure_probabilities",
+    "NumpyWavefunctionSimulator",
+    "ReferenceDensitySimulator",
+    "ReferenceWavefunctionSimulator",
+    "targeted_einsum",
+    "targeted_tensordot",
+    "zero_state_matrix",
+]
+
+from pyquil.simulation._numpy import (
+    NumpyWavefunctionSimulator,
+    get_measure_probabilities,
+    targeted_einsum,
+    targeted_tensordot,
+)
+from pyquil.simulation._reference import (
+    ReferenceDensitySimulator,
+    ReferenceWavefunctionSimulator,
+    zero_state_matrix,
+)