+++ test/unit/test_rewrite_arithmetic.py
@@ -36,7 +36,7 @@ def test_rewrite_arithmetic_duplicate_exprs():
 
     assert response == RewriteArithmeticResponse(
         original_memory_descriptors={"theta": ParameterSpec(length=1, type="REAL")},
-        recalculation_table={ParameterAref(index=0, name="__P1"): "theta[0]*1.5/(2*pi)"},
+        recalculation_table={ParameterAref(index=0, name="__P1"): "theta[0]*(1.5)/(2*pi)"},
         quil=Program("DECLARE __P1 REAL[1]", "DECLARE theta REAL[1]", "RZ(__P1[0]) 0", "RX(__P1[0]) 0").out(),
     )
 
@@ -58,14 +58,13 @@ def test_rewrite_arithmetic_mixed():
         ParameterAref(index=1, name="__P2"): "(beta[0] + theta[0])/(2*pi)",
     }
     assert (
-        response.quil
-        == Program(
+        Program(response.quil) == Program(
             "DECLARE __P2 REAL[2]",
             "DECLARE theta REAL[1]",
             "DECLARE beta REAL[1]",
             "RZ(__P2[0]) 0",
             "RZ(__P2[1]) 0",
-        ).out()
+        )
     )